backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
aiosqlite==0.20.0
beautifulsoup4==4.12.3
feedparser==6.0.11
apscheduler==3.10.4
python-dotenv==1.0.1
PyJWT==2.8.0
pydantic==2.6.4
lxml==5.1.0
